Basic Java Calculator using Java Forms
Cal1 Calculator is a simple calculator that perform arithmetic operations ‘+’, ‘-‘, ‘/’, ‘*’ & ’%’ based on user input, sums the up with an equals and displays the answer in a text box.
The calculator has been created using a specific set of global variables to complete all of the calculations required, which are:
	double doubleone;
	double doubletwo;
	char msoperator;
	int decimalClick;
The variables doubleone and doubletwo store the first and second number entered into the calculator. These have been create as doubles to store decimals values which cannot be achieved with interger, also to get a more accurate calculation instead in a rounded up number.
The char msoperator is used for my switch statements located in the equals button and is a char as it takes the characters from the functions of the calculator ‘+’, ‘-‘, ‘/’, ‘*’ & ’%’ and can be used to make multiple calculations
Int decimalClick is a simple operation to ensure that the ‘.’ decimal point cannot be pressed more than once when using the calculator, the decimalClick has been placed under each character and set to 0.
The calculator performs other functions such as:
Clear (C): This sets both double variables to zero and clears the main text box Tf1 to that it is empty
Equals (=): This contains the switch msoperator and sums up all of the calculations.